requirements.txt
# Core dependencies
lyricsgenius>=3.0.1
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.0
pydantic>=2.0.0
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0

# Optional: Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
